requirements.txt
aiogram
asyncio
loguru
openai
pandas
python-dotenv
SQLAlchemy
alembic
fastapi
uvicorn
pydantic[email]
jinja2
starlette
py_csv_xls
